# go/api/openapi.json
{
    "components": {
        "schemas": {
...
            "ATProtoStatusOutputBody": {
                "additionalProperties": false,
                "properties": {
                    "$schema": {
                        "description": "A URL to the JSON Schema for this object.",
                        "examples": [
                            "https://example.com/schemas/ATProtoStatusOutputBody.json"
                        ],
                        "format": "uri",
                        "readOnly": true,
                        "type": "string"
                    },
                    "connected": {
                        "type": "boolean"
                    },
                    "did": {
                        "type": "string"
                    },
                    "handle": {
                        "type": "string"
                    }
                },
                "required": [
                    "connected"
                ],
                "type": "object"
            },
...
            "ErrorDetail": {
                "additionalProperties": false,
                "properties": {
                    "location": {
                        "description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
                        "type": "string"
                    },
                    "message": {
                        "description": "Error message text",
                        "type": "string"
                    },
                    "value": {
                        "description": "The value at the given location"
                    }
                },
                "type": "object"
            },
            "ErrorModel": {
                "additionalProperties": false,
                "properties": {
                    "$schema": {
                        "description": "A URL to the JSON Schema for this object.",
                        "examples": [
                            "https://example.com/schemas/ErrorModel.json"
                        ],
                        "format": "uri",
                        "readOnly": true,
                        "type": "string"
                    },
                    "detail": {
                        "description": "A human-readable explanation specific to this occurrence of the problem.",
                        "examples": [
                            "Property foo is required but is missing."
                        ],
                        "type": "string"
                    },
                    "errors": {
                        "description": "Optional list of individual error details",
                        "items": {
                            "$ref": "#/components/schemas/ErrorDetail"
                        },
                        "type": [
                            "array",
                            "null"
                        ]
                    },
                    "instance": {
                        "description": "A URI reference that identifies the specific occurrence of the problem.",
                        "examples": [
                            "https://example.com/error-log/abc123"
                        ],
                        "format": "uri",
                        "type": "string"
                    },
                    "status": {
                        "description": "HTTP status code",
                        "examples": [
                            400
                        ],
                        "format": "int64",
                        "type": "integer"
                    },
                    "title": {
                        "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
                        "examples": [
                            "Bad Request"
                        ],
                        "type": "string"
                    },
                    "type": {
                        "default": "about:blank",
                        "description": "A URI reference to human-readable documentation for the error.",
                        "examples": [
                            "https://example.com/errors/example"
                        ],
                        "format": "uri",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "title": "Sunred API",
        "version": "1.0.0"
    },
    "openapi": "3.1.0",
    "paths": {
...
        "/api/v1/me/atproto": {
            "get": {
                "operationId": "atproto-status",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ATProtoStatusOutputBody"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "default": {
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        },
                        "description": "Error"
                    }
                },
                "summary": "Get AT Proto identity for the current user",
                "tags": [
                    "social"
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": ""
        }
    ],
...
}